Update agent.py
The Agent Class has been extended by including new attributes and methods which allows the agents to know which is the CS target (just in case the UAV decide to go to CS) and to manage better the battery consumptions due to each move. Also some minor changes have been made.