Update scenario_objects.py
Now it is possible to create different users clusters. The problem related to the different order of the centroids of the clusters (which occured often during the cluster detection after their generation) has been solved. Some minor changes have been made to properly adapt the environment to both 2D and 3D cases.