Update env_wrapper.py
Everything has been modified in order to manage the changes made in all other files. In particular now is possible to see on terminal (and to save automatically) all the parameters setting related to the environment and to the training. It is possible to visualize them (on terminal) better by uncomment the first 'time.sleep(5)' which appears inside the code. Some mismatch in folders error is possible since almost everything is saved at the end of the training and I did not check every single test case.